requirements.txt
Flask
tensorflow
numpy
pandas
Pillow
scikit-learn
imblearn
matplotlib
seaborn
